RF: Move ManagedDeviceError to exceptions
